requirements.txt
fastmcp>=0.9.0
slack-sdk>=3.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0

